docs: 디렉토리 구조 정리 & enum 필드 대문자로 변경
