--- src/main/java/com/kuit/agarang/common/model/entity/Status.java
@@ -1,5 +1,0 @@
-package com.kuit.agarang.common.model.entity;
-
-public enum Status {
-  ACTIVE, INACTIVE
-}
