+++ src/main/java/com/kuit/agarang/domain/ai/dto/TypecastResponse.java
@@ -1,4 +1,4 @@
-package com.kuit.agarang.domain.memory.dto;
+package com.kuit.agarang.domain.ai.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
@@ -6,13 +6,13 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class TypecastAudioResponse {
+public class TypecastResponse {
   private Result result;
 
   @Getter
   @NoArgsConstructor
   public static class Result {
-    @JsonProperty("audio_download_url")
-    private String audioDownloadUrl;
+    @JsonProperty("speak_v2_url")
+    private String speakV2Url;
   }
 }
