+++ src/main/java/com/kuit/agarang/domain/member/model/entity/RefreshToken.java
@@ -1,6 +1,6 @@
 package com.kuit.agarang.domain.member.model.entity;
 
-import com.kuit.agarang.common.model.entity.BaseEntity;
+import com.kuit.agarang.global.entity.BaseEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
