+++ src/main/java/com/kuit/agarang/domain/memory/model/entity/MusicBookmark.java
@@ -1,6 +1,6 @@
 package com.kuit.agarang.domain.memory.model.entity;
 
-import com.kuit.agarang.common.model.entity.BaseEntity;
+import com.kuit.agarang.global.entity.BaseEntity;
 import com.kuit.agarang.domain.member.model.entity.Member;
 import jakarta.persistence.*;
 import lombok.*;
