--- src/main/java/com/kuit/agarang/domain/memory/model/entity/enums/Tempo.java
@@ -1,5 +1,0 @@
-package com.kuit.agarang.domain.memory.model.entity.enums;
-
-public enum Tempo {
-  Fast, MID, SLOW
-}
