+++ src/main/java/com/kuit/agarang/domain/memory/model/enums/Genre.java
@@ -1,4 +1,4 @@
-package com.kuit.agarang.domain.memory.model.entity.enums;
+package com.kuit.agarang.domain.memory.model.enums;
 
 public enum Genre {
   BALLAD, POP, JAZZ, ACOUSTIC, RNB, ELECTRONIC, ROCK, INDIE, HIPHOP
