+++ src/main/java/com/kuit/agarang/domain/memory/model/enums/Instrument.java
@@ -1,4 +1,4 @@
-package com.kuit.agarang.domain.memory.model.entity.enums;
+package com.kuit.agarang.domain.memory.model.enums;
 
 public enum Instrument {
   PIANO, VIOLIN, FLUTE, CLARINET, BASE_GUITAR, ELECTRIC_GUITAR,
