+++ src/main/java/com/kuit/agarang/domain/memory/model/enums/Mood.java
@@ -1,4 +1,4 @@
-package com.kuit.agarang.domain.memory.model.entity.enums;
+package com.kuit.agarang.domain.memory.model.enums;
 
 public enum Mood {
   BEAUTIFUL, BRIGHT, HAPPY, PEACEFUL, WARM, ENERGETIC, FANTASTIC, LOVELY, TOUCHING
