+++ src/main/java/com/kuit/agarang/domain/memory/model/enums/Tempo.java
@@ -1,0 +1,5 @@
+package com.kuit.agarang.domain.memory.model.enums;
+
+public enum Tempo {
+  FAST, MID, SLOW
+}
