+++ src/main/java/com/kuit/agarang/domain/playlist/model/entity/Playlist.java
@@ -1,12 +1,11 @@
 package com.kuit.agarang.domain.playlist.model.entity;
 
-import com.kuit.agarang.common.model.entity.BaseEntity;
+import com.kuit.agarang.global.entity.BaseEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import lombok.*;
-import lombok.experimental.SuperBuilder;
 
 @Entity
 @Getter
