+++ src/main/java/com/kuit/agarang/global/deploy/model/dto/ServerInfo.java
@@ -1,16 +1,20 @@
-package com.kuit.agarang.global.deploy.dto;
+package com.kuit.agarang.global.deploy.model.dto;
 
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
-@Builder
-@AllArgsConstructor
 @NoArgsConstructor
 public class ServerInfo {
   private String port;
   private String address;
   private String name;
+
+  @Builder
+  public ServerInfo(String port, String address, String name) {
+    this.port = port;
+    this.address = address;
+    this.name = name;
+  }
 }
