+++ src/main/java/com/kuit/agarang/global/entity/Status.java
@@ -1,0 +1,5 @@
+package com.kuit.agarang.global.entity;
+
+public enum Status {
+  ACTIVE, INACTIVE
+}
